.tmux.conf: Add basename of pane's path to the status  line
Background

Some times the my tmux contains too much panes. To the easy identification
the status bar should provide some information about the working path of panes.

Decision

Let's add pane's path as an identifier to the status line. But the full path
sometimes long and it makes ugly the status bar.

Restrict the path to the last element of the path (a.k.a.) basename. With this
basename some useful information is provided from the pane's working directory.

Consequences
- the status line shows the basename of current working directory of the pane.